pyproject.toml
# Copyright (C) 2020-2026 Serghei Iakovlev <gnu@serghei.pl>
#
# This file is free software; you can redistribute it and/or
# modify it under the terms of the GNU General Public License
# as published by the Free Software Foundation; either version 3
# of the License, or (at your option) any later version.
#
# This file is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this file.  If not, see <https://www.gnu.org/licenses/>.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gstore"
version = "0.8.0"
description = "Synchronize GitHub repositories made easy."
readme = "README.rst"
license = "GPL-3.0-or-later"
requires-python = ">=3.9"
authors = [{ name = "Serghei Iakovlev", email = "gnu@serghei.pl" }]
maintainers = [{ name = "Serghei Iakovlev", email = "gnu@serghei.pl" }]
keywords = ["git", "github", "backup", "repo", "sync", "cli"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "Natural Language :: English",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: System :: Archiving :: Backup",
    "Topic :: System :: Software Distribution",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Version Control",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = ["pygithub>=2.5.0,<3", "gitpython>=3.1.31,<4"]

[project.optional-dependencies]
testing = [
    "coverage[toml]>=7.2.2,<8",
    "pytest>=8.3.3,<9",
    "pytest-mock>=3.10.0,<4",
    "ruff>=0.9.0",
]
docs = ["sphinx>=6.1.3", "furo>=2023.3.27"]
build = ["twine>=6.0.0,<7"]
dev = ["debugpy>=1.8.9,<2"]

[project.scripts]
gstore = "gstore.cli:main"

[project.urls]
Homepage = "https://gstore.readthedocs.io"
Documentation = "https://gstore.readthedocs.io"
Repository = "https://github.com/sergeyklay/gstore"
"Bug Tracker" = "https://github.com/sergeyklay/gstore/issues"
Changelog = "https://gstore.readthedocs.io/en/latest/changelog.html"

[dependency-groups]
dev = [
    "coverage[toml]>=7.2.2,<8",
    "pytest>=8.3.3,<9",
    "pytest-mock>=3.10.0,<4",
    "ruff>=0.9.0",
    "sphinx>=6.1.3",
    "furo>=2023.3.27",
    "twine>=6.0.0,<7",
    "debugpy>=1.8.9,<2",
]

[tool.hatch.build.targets.sdist]
include = [
    "gstore/",
    "tests/",
    "AUTHORS.rst",
    "CHANGELOG.rst",
    "CONTRIBUTING.rst",
    "LICENSE",
    "MAINTAINERS.rst",
    "README.rst",
    "SECURITY.rst",
]

[tool.hatch.build.targets.wheel]
packages = ["gstore"]

[tool.pytest.ini_options]
addopts = "--verbose --doctest-modules --durations=25"

[tool.ruff]
target-version = "py39"
line-length = 79
extend-exclude = [".venv", "__pycache__", "dist"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "BLE", # flake8-blind-except
    "C90", # McCabe complexity
    "PLE", # Pylint errors
    "PLW", # Pylint warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]
ignore = [
    "PLW0603", # Global statement for _proc_ctx in repo.py (required for multiprocessing)
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["gstore"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = []

[tool.coverage.report]
# Set the number of decimal places in coverage percentage
precision = 2

# Show lines that were not executed
show_missing = true

[tool.coverage.html]
# Directory for the HTML format report
directory = "htmlcov"

[tool.coverage.xml]
# Filename for the XML format report
output = "coverage.xml"

[tool.coverage.run]
# Enable branch coverage (default: False)
branch = true

# Run tests in parallel (useful when using CI/CD)
parallel = true

# Filename for storing coverage data
data_file = ".coverage"

# Omit specified files or directories from the report
omit = ["*/.venv/*", "*/__pycache__/*", "*/site-packages/*", "*/tests/*"]
